requirements.txt
docopt==0.6.2
dulwich==0.20.24
gitpython==3.1.41
pylint==2.11.1
requests==2.32.4
pytest==6.2.5
black==24.3.0
flake8==4.0.1
zipp>=3.19.1 # not directly required, pinned by Snyk to avoid a vulnerability
setuptools>=78.1.1 # not directly required, pinned by Snyk to avoid a vulnerability
urllib3>=2.5.0 # not directly required, pinned by Snyk to avoid a vulnerability
